Move classes of the old bootstrapping to their own package
