+++ lib/src/main/java/org/sonarsource/scanner/lib/internal/facade/inprocess/Slf4jLogOutputAdapter.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.scanner.lib.internal;
+package org.sonarsource.scanner.lib.internal.facade.inprocess;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
